LPS-123192 Move plugins
