+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/CSSBuilderDefaultsPlugin.java
@@ -12,9 +12,8 @@
  * details.
  */
 
-package com.liferay.gradle.plugins.internal;
+package com.liferay.gradle.plugins;
 
-import com.liferay.gradle.plugins.BaseDefaultsPlugin;
 import com.liferay.gradle.plugins.css.builder.BuildCSSTask;
 import com.liferay.gradle.plugins.css.builder.CSSBuilderPlugin;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
