+++ modules/sdk/gradle-plugins/src/main/resources/META-INF/gradle-plugins/com.liferay.idea.defaults.properties
@@ -1,0 +1,1 @@
+implementation-class=com.liferay.gradle.plugins.IdeaDefaultsPlugin
